+++ packaging/test_installed_package.py
@@ -19,8 +19,8 @@ TEST_TIMEOUT_SECONDS = 5 * 60
 DEPENDENCIES = [
     ("mojo-net", "0.2.2", PACKAGES / "mojo-net"),
     ("protomojo", "0.3.0", PACKAGES / "protomojo"),
-    ("mojo-tls", "0.2.4", PACKAGES / "mojo-tls"),
-    ("mojo-http2", "0.2.4", PACKAGES / "mojo-http2"),
+    ("mojo-tls", "0.3.0", PACKAGES / "mojo-tls"),
+    ("mojo-http2", "0.2.5", PACKAGES / "mojo-http2"),
 ]